Added a classic way for the client unread count.
